Cleanup bootstrap / move config stuff to config.py
